+++ SharpSource/SharpSource.Test/AsyncOverloadsAvailableTests.cs
@@ -1007,4 +1007,37 @@ async void MyMethod()
 }";
         await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("Async overload available for StringWriter.Write"), result);
     }
+
+    [BugVerificationTest(IssueUrl = "https://github.com/Vannevelj/SharpSource/issues/295")]
+    public async Task AsyncOverloadsAvailable_MaintainsIndentation()
+    {
+        var original = @"
+using System.Linq;
+using System.IO;
+
+async void MyMethod()
+{
+    // A comment
+
+    var text = {|#0:File.ReadAllText(""file.txt"")|}
+    .Trim()
+    .Split('|')
+    .ToList();
+}";
+
+        var result = @"
+using System.Linq;
+using System.IO;
+
+async void MyMethod()
+{
+    // A comment
+
+    var text = {|#0:(await File.ReadAllTextAsync(""file.txt""))|}
+    .Trim()
+    .Split('|')
+    .ToList();
+}";
+        await VerifyCS.VerifyCodeFix(original, VerifyCS.Diagnostic().WithMessage("Async overload available for File.ReadAllText"), result);
+    }
 }